Move the panel update login to app.tsx
